Applied requested changes, improved application test, changed imp.bidFloor from Float to BigDecimal and fixed corresponding errors

+++ src/main/java/org/prebid/server/bidder/appnexus/AppnexusBidder.java
@@ -136,7 +136,7 @@ public class AppnexusBidder implements Bidder<BidRequest> {
 
         final BigDecimal reserve = appnexusExt.getReserve();
         if (reserve != null && reserve.compareTo(BigDecimal.ZERO) > 0) {
-            impBuilder.bidfloor(reserve.floatValue()); // This will be broken for non-USD currency.
+            impBuilder.bidfloor(reserve); // This will be broken for non-USD currency.
         }
 
         return ImpWithMemberId.of(impBuilder.build(), appnexusExt.getMember());
